Fixed shadowed declaration and minor cleanup
I fixed that issue with the idx being declared twice and checked to make sure it compiles. As well, there was a comment mentioning to use std::partition, but that is already done just the line below it, so I think that comment was old. 

If the trace_options_.cull_back_face section looks odd, this might help. Previously, it was of the form: (A && B) || (B && C). That simplifies to B && (A || C).